split interfact/impl
